kref: remove WARN_ON for NULL release functions
The kref functions check for NULL release functions.  This WARN_ON seems
rather pointless.  We will eventually release and then just crash nicely. 
It is also somewhat expensive because these functions are inlined in a lot
of places.  Removing the WARN_ONs saves around 2.3k in this kernel (likely
more in others with more drivers)

   text    data     bss     dec     hex filename
9083992 5367600 11116544        25568136        1862388 vmlinux-before-load-avg
9070166 5367600 11116544        25554310        185ed86 vmlinux-load-avg

Link: http://lkml.kernel.org/r/20170315021431.13107-5-andi@firstfloor.org
Signed-off-by: Andi Kleen <ak@linux.intel.com>
Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>