net: use kvmalloc with __GFP_REPEAT rather than open coded variant
fq_alloc_node, alloc_netdev_mqs and netif_alloc* open code kmalloc with
vmalloc fallback.  Use the kvmalloc variant instead.  Keep the
__GFP_REPEAT flag based on explanation from Eric:

"At the time, tests on the hardware I had in my labs showed that vmalloc()
could deliver pages spread all over the memory and that was a small
penalty (once memory is fragmented enough, not at boot time)"

The way how the code is constructed means, however, that we prefer to
go and hit the OOM killer before we fall back to the vmalloc for
requests <=32kB (with 4kB pages) in the current code.  This is rather
disruptive for something that can be achived with the fallback.  On the
other hand __GFP_REPEAT doesn't have any useful semantic for these
requests.  So the effect of this patch is that requests which fit into
32kB will fall back to vmalloc easier now.

Link: http://lkml.kernel.org/r/20170306103327.2766-3-mhocko@kernel.org
Signed-off-by: Michal Hocko <mhocko@suse.com>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Eric Dumazet <edumazet@google.com>
Cc: David Miller <davem@davemloft.net>
Signed-off-by: Andrew Morton <>