zram: handle multiple pages attached to bio's bvec
Johannes Thumshirn reported system goes the panic when using NVMe over
Fabrics loopback target with zram.

The reason is zram expects each bvec in bio contains a single page but
nvme can attach a huge bulk of pages attached to the bio's bvec so that
zram's index arithmetic could be wrong so that out-of-bound access makes
panic.

It was solved by limiting max_sectors with SECTORS_PER_PAGE in
0bc315381fe9 ("zram: set physical queue limits to avoid array out of
bounds accesses") but that makes zram slow because a bio should split
with each pages.  So this patch makes zram aware of multiple pages in a
bvec so it can solve the panic without causing any performance
regression.

Link: http://lkml.kernel.org/r/1491196653-7388-2-git-send-email-minchan@kernel.org
Signed-off-by: Johannes Thumshirn <jthumshirn@suse.de>
Signed-off-by: Minchan Kim <minchan@kernel.org>
Reported-by: Johannes Thumshirn <jthumshirn@suse.de>
Tested-by: Johannes Thumshirn <jthumshirn@suse.de>
Reviewed-by: Johannes Thumshirn <jthumshirn@suse.de>
Reviewed-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: Jens Axboe <axboe@kernel.dk>
Cc: Hannes Reinecke <hare@suse.com>
Cc: Mika Penttil <mika.penttila@nextfour.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>