checkpatch: allow space leading blank lines in email headers
Allow a leading space and otherwise blank link in the email headers as it
can be a line wrapped Spamassassin multiple line string or any other valid
rfc 2822/5322 email header.

The line with space causes checkpatch to erroniously think that it's in
the content body, as opposed to headers and thus flag a mail header as an
unwrapped long comment line.

Link: http://lkml.kernel.org/r/d75a9f0b78b3488078429f4037d9fff3bdfa3b78.1490247180.git.joe@perches.com
Signed-off-by: Joe Perches <joe@perches.com>Reported-by: Darren Hart (VMware) <dvhart@infradead.org>
Tested-by: Darren Hart (VMware) <dvhart@infradead.org>
Reviewed-by: Darren Hart (VMware) <dvhart@vmware.com>
Original-patch-by: John 'Warthog9' Hawley (VMware) <warthog9@eaglescrag.net>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>