mm, swap: sort swap entries before free
To reduce the lock contention of swap_info_struct->lock when freeing swap
entry.  The freed swap entries will be collected in a per-CPU buffer
firstly, and be really freed later in batch.  During the batch freeing, if
the consecutive swap entries in the per-CPU buffer belongs to same swap
device, the swap_info_struct->lock needs to be acquired/released only
once, so that the lock contention could be reduced greatly.  But if there
are multiple swap devices, it is possible that the lock may be
unnecessarily released/acquired because the swap entries belong to the
same swap device are non-consecutive in the per-CPU buffer.

To solve the issue, the per-CPU buffer is sorted according to the swap
device before freeing the swap entries.  Test shows that the time spent by
swapcache_free_entries() could be reduced after the patch.

The patch was tested by measuring the run time of
swap_cache_free_entries() during the exit phase of an application which
uses much swap space.  The results show that the average run time of
swap_cache_free_entries() was reduced about 20% after applying the patch.

The average run time of swap_cache_free_entries() is reduced from about
~137us to ~111us.  The total samples of swap_cache_free_entries() is
about 200000, run on 16 CPUs, so the wall time is about 1.7s.  I will
revise the tests to get the total run time reduction.

I added memory freeing timing capability to vm-scalability test suite. 
The result shows the memory freeing time reduced from 2.64s to 2.31s
(about -12.5%).  The memory space to free is 96G (including swap).  The
machine has 144 CPU, 32G RAM, and 96G swap.  The process number is 16.

Link: http://lkml.kernel.org/r/20170407064901.25398-1-ying.huang@intel.com
Signed-off-by: Huang Ying <ying.huang@intel.com>
Acked-by: Tim Chen <tim.c.chen@intel.com>
Acked-by: Rik van Riel <riel@redhat.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Shaohua Li <shli@kernel.org>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>