test_sysctl: add dedicated proc sysctl test driver
Patch series "sysctl: add dedicated stress test driver".

This series adds a dedicated proc sysctl stress test driver and extends
our current set of tests.  I spun this off from a larger series to first
get the basic knobs in place for unsigned int, which are now merged in
your tree, so this series depends on those patches.

The stress test driver uses a new license (GPL on Linux, copyleft-next
outside of Linux).  Linus was fine with this [0] and later due to Ted's
and Alans's request ironed out an "or" language clause to use [1] which is
already present upstream.

[0] https://lkml.kernel.org/r/CA+55aFyhxcvD+q7tp+-yrSFDKfR0mOHgyEAe=f_94aKLsOu0Og@mail.gmail.com
[1] https://lkml.kernel.org/r/1495234558.7848.122.camel@linux.intel.com



This patch (of 6):

Although we have had tools/testing/selftests/sysctl/ with two test cases
these use existing kernel sysctl interfaces.  We want to expand test
coverage, we can't just be looking for random safe production values to
poke at, instead just dedicate a test driver for debugging purposes and
port the existing scripts to use it.  This will make it easier for further
tests to be added.

Link: http://lkml.kernel.org/r/20170525233729.18530-2-mcgrof@kernel.org
Signed-off-by: Luis R. Rodriguez <mcgrof@kernel.org>
Acked-by: Kees Cook <keescook@chromium.org>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Shuah Khan <shuah@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>