zram: compare all the entries with same checksum for deduplication
Until now, we compare just one entry with same checksum when checking
duplication since it is the simplest way to implement.  However, for the
completeness, checking all the entries is better so this patch implement
to compare all the entries with same checksum.  Since this event would be
rare so there would be no performance loss.

Link: http://lkml.kernel.org/r/1494556204-25796-5-git-send-email-iamjoonsoo.kim@lge.com
Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Reviewed-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Acked-by: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>