kernel/watchdog_hld.c: fix spurious hard lockups
Some users reported spurious NMI watchdog timeouts.

We now have more and more systems where the Turbo range is wide enough
that the NMI watchdog expires faster than the soft watchdog timer that
updates the interrupt tick the NMI watchdog relies on.

This problem was originally added by 58687acba592 ("lockup_detector:
Combine nmi_watchdog and softlockup detector").  Previously the NMI
watchdog would always check jiffies, which were ticking fast enough.  But
now the backing is quite slow so the expire time becomes more sensitive.

For mainline the right fix is to switch the NMI watchdog to reference
cycles, which tick always at the same rate independent of turbo mode.  But
this is requires some complicated changes in perf, which are too difficult
to backport.  Since we need a stable fix too just increase the NMI
watchdog rate here to avoid the spurious timeouts.  This is not an ideal
fix because a 3x as large Turbo range could still fail, but for now that's
not likely.

The right fix for mainline can be found here.
perf/x86/intel: enable CPU ref_cycles for GP counter
perf/x86/intel, watchdog: Switch NMI watchdog to ref cycles on x86
https://patchwork.kernel.org/patch/9779087/
https://patchwork.kernel.org/patch/9779089/

Link: http://lkml.kernel.org/r/20170620213309.30051-1-kan.liang@intel.com
Fixes: 58687acba592 ("lockup_detector: Combine nmi_watchdog and softlockup detector")
Signed-off-by: Kan Liang <Kan.liang@intel.com>
Cc: Don Zickus <dzickus@redhat.com>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Babu Moger <babu.moger@oracle.com>
Cc: Aaron Tomlin <atomlin@redhat.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Stephane Eranian <eranian@google.com>
Cc: Arnaldo Carvalho de Melo <acme@redhat.com>
Cc: Andi Kleen <ak@linux.intel.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>