checkpatch: change format of --color argument to --color[=WHEN]
The boolean --color argument did not offer the ability to force colourized
output even if stdout is not a terminal.  Change the format of the
argument to the familiar --color[=WHEN] construct as seen in common Linux
utilities such as git, ls and dmesg, which allows the user to specify
whether to colourize output "always", "never", or "auto" when the output
is a terminal.  The default is "auto".

The old command-line uses of --color and --no-color are unchanged.

Link: http://lkml.kernel.org/r/efe43bdbad400f39ba691ae663044462493b0773.1496799721.git.joe@perches.com
Signed-off-by: John Brooks <john@fastquake.com>
Signed-off-by: Joe Perches <joe@perches.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>